requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
albucore==0.0.13
albumentations==1.4.10
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
astor==0.8.1
astunparse==1.6.3
attrs==24.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.11
datasets==2.10.0
decorator==5.1.1
dill==0.3.6
filelock==3.17.0
fire==0.7.0
flatbuffers==25.1.21
fonttools==4.55.4
frozenlist==1.5.0
fsspec==2024.9.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
grpcio==1.69.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
idna==3.10
imageio==2.37.0
imgaug==0.4.0
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.8.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
lmdb==1.6.2
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.14
namex==0.0.8
narwhals==1.23.0
networkx==3.4.2
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt-einsum==3.3.0
optree==0.14.0
packaging==24.2
paddleocr==2.9.1
paddlepaddle==2.6.2
pandas==2.2.3
pillow==11.1.0
propcache==0.2.1
protobuf==3.20.2
psutil==6.1.1
py-cpuinfo==9.0.0
pyarrow==19.0.0
pyclipper==1.3.0.post6
pydantic==2.10.5
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.11.0
referencing==0.36.1
regex==2024.11.6
requests==2.32.3
responses==0.18.0
rich==13.9.4
rpds-py==0.22.3
safetensors==0.5.2
scikit-image==0.25.0
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
seqeval==1.2.2
setuptools==75.8.0
shapely==2.0.6
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
streamlit==1.41.1
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2025.1.10
tokenizers==0.21.0
toml==0.10.2
tomli==2.2.1
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.48.1
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.65
ultralytics-thop==2.0.14
urllib3==2.3.0
watchdog==6.0.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3